requirements.txt
# Core trading dependencies
MetaTrader5==5.0.5120
pandas==2.3.1
numpy==2.3.1
matplotlib==3.10.3
ta==0.11.0
schedule==1.2.2
optuna==4.4.0
python-dotenv==1.1.1
scikit-learn==1.7.1

# Data analysis and visualization
seaborn==0.12.2
plotly==5.15.0

# Jupyter notebook support
jupyter==1.1.1
notebook==7.4.4
ipykernel==6.29.5

# Additional utilities
requests==2.32.4
PyYAML==6.0.2

# Testing (optional)
pytest==7.4.0
pytest-cov==4.1.0
